event/iot_service/requirements.txt
psycopg2-binary==2.9.3
pika==1.2.0
